+++ src/components/dialogs/types.ts
@@ -1,0 +1,12 @@
+export interface DialogBaseProps {
+  open?: boolean;
+  children?: React.ReactNode;
+  onClose?: () => void;
+  onAccept?: () => void;
+  disableFooter?: boolean;
+  title?: string;
+  labels?: {
+    accept?: string;
+    cancel?: string;
+  };
+}