+++ src/components/index.ts
@@ -1,4 +1,5 @@
 import AppWrapper from "./app-wrapper";
+import Icon from "./icon";
 import Menu from "./menu";
 
-export { AppWrapper, Menu };
+export { AppWrapper, Icon, Menu };