+++ src/components/learning/courses-tracking/courses/CoursesList.tsx
@@ -1,16 +1,53 @@
 import { useGet } from "contexts/api-context/hooks";
 import { CourseSchema } from "types/schemas/courses";
-import CourseItem from "./CourseItem";
+import { useState } from "react";
+import Dialog from "components/dialogs/dialog";
+import AddCourse from "./AddCourse";
+import { Button } from "components/buttons";
+import Table from "components/table/Table";
+import { useRouting } from "hooks";
 
 const CoursesList: React.FC = () => {
-  const [courses, loading] = useGet<CourseSchema[]>("courses.list");
-  console.log(courses);
+  const [courses, loading, { refetch }] =
+    useGet<CourseSchema[]>("courses.list");
+  const { goTo } = useRouting();
+  const [openedAdd, setOpenedAdd] = useState(false);
+  const toggleAdd = () => setOpenedAdd(!openedAdd);
+  const handleSaved = () => {
+    refetch();
+    toggleAdd();
+  };
   return (
     <>
+      <div>
+        <Button onClick={toggleAdd}>Add</Button>
+        <Button onClick={() => goTo("learningFollowUp")}>Add follow up</Button>
+      </div>
       {loading && <p>Loading...</p>}
-      {courses?.map((course) => (
+      <Table
+        title="List of courses"
+        colLabels={{
+          title: "Title",
+          description: "Description",
+          completed_lessons: "Completed",
+          percentage: "Progress",
+          lessons: "Lessons",
+        }}
+        columns={[
+          "title",
+          "description",
+          "lessons",
+          "percentage",
+          "completed_lessons",
+        ]}
+        data={courses}
+      />
+      {/* {courses?.map((course) => (
         <CourseItem course={course} key={`course-${course.id}`} />
-      ))}
+      ))} */}
+      <Dialog open={openedAdd} title={"Some cool"} disableFooter>
+        <AddCourse onCancel={toggleAdd} onSaved={handleSaved} />
+      </Dialog>
     </>
   );
 };