+++ src/components/menu/Menu.tsx
@@ -1,3 +1,4 @@
+import Icon from "components/icon";
 import { NavLink } from "react-router-dom";
 
 const Menu: React.FC = () => {
@@ -6,15 +7,19 @@ const Menu: React.FC = () => {
       <div id="main-side-bar">
         <ul>
           <li>
+            <Icon icon="gauge" />
             <NavLink to="/dashboard">Dashboard</NavLink>
           </li>
           <li>
+            <Icon icon="graduation-cap" />
             <NavLink to="/learning">Learning</NavLink>
           </li>
           <li>
+            <Icon icon="cogs" />
             <NavLink to="/settings">Settings</NavLink>
           </li>
           <li>
+            <Icon icon="user" />
             <NavLink to="/profile">Profile</NavLink>
           </li>
         </ul>