+++ src/hooks/use-prev-props.ts
@@ -1,0 +1,10 @@
+import { useRef, useEffect } from "react";
+const usePrevProps = (value: any) => {
+  const valuesRef = useRef(value);
+  useEffect(() => {
+    valuesRef.current = value;
+  });
+  return valuesRef.current;
+};
+
+export default usePrevProps;