+++ src/screens/authenticated/index.ts
@@ -2,5 +2,12 @@ import DashboardScreen from "./dashboard-screen";
 import LearningScreen from "./learning-screen/LearningScreen";
 import ProfileScreen from "./profile-screen";
 import SettingsScreen from "./settings-screen";
+import LearningFollowUpScreen from "./learning-followup-screen";
 
-export { DashboardScreen, LearningScreen, ProfileScreen, SettingsScreen };
+export {
+  DashboardScreen,
+  LearningScreen,
+  LearningFollowUpScreen,
+  ProfileScreen,
+  SettingsScreen,
+};