+++ src/screens/authenticated/learning-followup-screen/LearningFollowupScreen.tsx
@@ -1,0 +1,9 @@
+const LearningFollowUpScreen: ScreenType = () => {
+  return (
+    <div>
+      <h3>Learning follow up</h3>
+    </div>
+  );
+};
+
+export default LearningFollowUpScreen;