+++ src/utils/registryIcons.ts
@@ -1,0 +1,10 @@
+import { library } from "@fortawesome/fontawesome-svg-core";
+import {
+  faTag,
+  faGauge,
+  faGraduationCap,
+  faCogs,
+  faUser,
+} from "@fortawesome/free-solid-svg-icons";
+
+library.add(faCogs, faGauge, faGraduationCap, faTag, faUser);